move addResult() to seperate file
